Implement credit admin functionality for support and analysis
The credit model is now more relevant than transactions for
examination of data, and as such much of the functionality added to
the transaction admin has been moved to the credit admin.
